requirements/base.txt
asgiref==3.2.10
attrs==20.2.0
Django==3.1.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.1
drf-nested-routers==0.92.1
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
